pyproject.toml
[project]
name = "nexus-erp"
version = "0.1.0"
description = "Nexus ERP - Sistema de Gestión Médica e Inteligencia de Negocios"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pyjwt>=2.11.0",
    "sqlalchemy>=2.0.47",
    "uvicorn>=0.41.0",
    "websockets>=16.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
]